use __builtin_alloca, drop non-standard alloca.h
Since alloca.h is not ISO C, most of our alloca-using tests seem to
rely on __builtin_alloca instead of including the header and calling
alloca.  This patch extends this practice to some of the exceptions I
found in gcc.target, marking them as requiring a functional alloca
while at that.


for  gcc/testsuite/ChangeLog

	* gcc.target/arc/interrupt-6.c: Use __builtin_alloca, require
	effective target support for alloca, drop include of alloca.h.
	* gcc.target/i386/pr80969-3.c: Likewise.
	* gcc.target/sparc/setjmp-1.c: Likewise.
	* gcc.target/x86_64/abi/ms-sysv/gen.cc: Likewise.
	* gcc.target/x86_64/abi/ms-sysv/ms-sysv.c: Likewise.

git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@274558 138bc75d-0d04-0410-961f-82ee72b054a4